requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.31.1
pydantic==2.9.2
pyyaml==6.0.1
requests==2.32.3

# AI/ML dependencies
ollama==0.3.3
chromadb==0.5.5
langchain==0.3.1
langchain-community==0.3.1
langchain-text-splitters==0.3.1
transformers>=4.44.2
torch>=2.8.0
safetensors>=0.4.5
accelerate>=0.34.2
soundfile>=0.12.1

# UI dependencies
streamlit==1.39.0

# Optional voice dependencies
gtts==2.5.3

# Development and testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
black==24.8.0
flake8==7.1.1

# Document processing
pypdf2==3.0.1
python-multipart==0.0.12
